KotbarView: split KotbarReservationCard into subcomponents
